Fix: SignUpDto 생성 후 request/response dto 생성, 회원가입 시 AuthController에서 UserEntity가 아닌 SignUpResponse를 반환하도록 수정
